chore: split into files and functions
